Rename lift.py to lift_function.py
